fix(simulator): a requantization shift of 31 overflowed the rounding divide's mask
D-0068. The shift is legal in [0, 31]: the format's check accepts 31 and
Section 14's decomposition emits it for every multiplier in [2^-32, 2^-31).
At 31 the kernel's mask, (int32_t{1} << 31) - 1, is INT32_MIN minus one, a
signed overflow. GCC wraps it to the right value, so the answer was correct
and the program was undefined.

Nothing executed that shift. The format's tests drive 32 and -1, the values
just outside the range, EncodingTest round trips 31 through the bytes without
running it, and every semantics case uses a shift between 0 and 7. It was found
by reading the arithmetic the QDQ contraction is about to reach, not by a
failure.

Proven both ways before the fix went in, in a throwaway UBSan build outside the
repository: at the parent's kernel the new case halts on "signed integer
overflow: -2147483648 - 1", and at the fix the whole binary passes, 79 tests.
The CI sanitizer job runs this binary under UBSan with halt_on_error, so the
case keeps the shift defined there from now on.

The mask, remainder and threshold are 64 bits, as gemmlowp builds them. Nothing
moves at shifts 0 to 30, and at 31 the value is the one the wrapped mask gave.
The case holds both columns at the int32 extremes and answers 1 and -1, which
the numpy reference answers too.

+++ lib/Simulator/Kernels.cpp
@@ -301,12 +301,18 @@ int32_t saturatingRoundingDoublingHighMul(int32_t a, int32_t b) {
 /// shift rather than as a division: a division would round toward zero and the
 /// correction term below is written for a shift that rounds toward negative
 /// infinity.
+///
+/// **The mask is 64 bits wide, and that is D-0068.** The shift is legal up to
+/// 31, and at 31 the mask `2^31 - 1` is one more than an int32 subtraction can
+/// produce from `1 << 31`: the 32 bit form was a signed overflow, undefined
+/// behaviour that happened to give the right bits on this compiler. gemmlowp
+/// builds the mask from a 64 bit one for the same reason.
 int32_t roundingDivideByPOT(int32_t value, int32_t exponent) {
   if (exponent <= 0)
     return value;
-  const int32_t mask = (int32_t{1} << exponent) - 1;
-  const int32_t remainder = value & mask;
-  const int32_t threshold = (mask >> 1) + (value < 0 ? 1 : 0);
+  const int64_t mask = (int64_t{1} << exponent) - 1;
+  const int64_t remainder = static_cast<int64_t>(value) & mask;
+  const int64_t threshold = (mask >> 1) + (value < 0 ? 1 : 0);
   return (value >> exponent) + (remainder > threshold ? 1 : 0);
 }
